AudioFlinger: Take resampling into account when choosing the minimum frame count
If the caller tried to use the fast path with a sample rate
different from the device's native sample rate (not knowing whether
audio flinger supports resampling in the fast path or not, which
is a compile time define), make sure to set the frame count to
a value high enough.

If resampling from a higher sample rate to a lower (e.g. playing
back 48 kHz content on a device with a native sample rate of 44.1
kHz), the input number of frames needs to be higher than the
output block size, otherwise the resampler will occasionally
reuse input samples (leading to small ticking sounds with regular
intervals, and triggering the "desiredFrames <= cblk->frameCount_"
assert if built with debugging).

This should fix playback of 48 kHz PCM content via OpenSL ES on
Nexus 7 (2012/grouper) - verified on the emulator which showed
the same issue as the Nexus 7.

Increasing the minimum number of buffers from 2 to 3 is similar
to what is done in AudioTrack::createTrack_l.

Fixes: http://b.android.com/59453

Change-Id: I3ef7ea32826da5105f16423a9f333025f5aa0add

Conflicts:
	services/audioflinger/Threads.cpp